% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{longest_word}
\alias{longest_word}
\title{Finds the longest words in a text}
\usage{
longest_word(x)
}
\arguments{
\item{x}{A string}
}
\value{
A character vector with the longest word and its length.
}
\description{
Words can be separated by a space, an apostrophe or by one of the following characters: \code{, . ; : ? !}
}
\details{
If there is more than one word with the max length, the first
in alphabetical order is returned
}
