% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{Counts the number of words in a text}
\usage{
word_count(x)
}
\arguments{
\item{x}{A string}
}
\value{
A numeric vector with the number of words
}
\description{
Words can be separated by a space, an apostrophe or by one of the following characters: \code{, . ; : ? !}
}
