% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_split}
\alias{word_split}
\title{Splits a string in a vector of words}
\usage{
word_split(x)
}
\arguments{
\item{x}{A string}
}
\value{
A character vector
}
\description{
Words can be separated by a space, an apostrophe or by one of the following characters: \code{, . ; : ? !}
}
